Extract budget related server handlers from main.ts to server/budget/app.ts
