+++ packages/loot-core/src/types/handlers.d.ts
@@ -1,6 +1,6 @@
 import type { AccountHandlers } from '../server/accounts/app';
 import type { AdminHandlers } from '../server/admin/types/handlers';
-import type { BudgetHandlers } from '../server/budget/types/handlers';
+import type { BudgetHandlers } from '../server/budget/app';
 import type { DashboardHandlers } from '../server/dashboard/types/handlers';
 import type { FiltersHandlers } from '../server/filters/types/handlers';
 import type { NotesHandlers } from '../server/notes/types/handlers';
